Fix opm authorization
This fixes an authorization problem occurring when
trying to render an OLM index container with opm.
There are two tools used in the precaching workload
from inside the container, podman and opm.
Podman is using auth.json from /run/containers/0
when run as root. Opm is using ~/.docker/config.json.
This change adds missing pull secret copy to the
latter for the opm use, and also puts all the relevant
copy operations to the "copy_env.sh" script
